fix: modify account modal styles
