+++ src/Components/Atoms/SocialLoginButton/index.ts
@@ -1,0 +1,3 @@
+import SocialLoginButtonViewer from './SocialLoginButtonViewer'
+
+export default SocialLoginButtonViewer
